move tests that's not snap to another file
