+++ packages/components/src/Slider/Slider.snap.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import Slider, { getSelectedIconPosition } from './Slider.component';
+import Slider from './Slider.component';
 
-describe('Slider component', () => {
+describe('Slider component snaps', () => {
 	describe('renderers', () => {
 		it('should render Slider', () => {
 			// given
@@ -51,26 +51,6 @@ describe('Slider component', () => {
 			const wrapper = shallow(<Slider {...props} />);
 			// then
 			expect(wrapper.getElement()).toMatchSnapshot();
-		});
-	});
-
-	describe('getSelectedIconPosition()', () => {
-		it('should return the selected position', () => {
-			// given
-			const icons = ['icon1', 'icon2', 'icon3', 'icon4', 'icon5'];
-			// when
-			const result = getSelectedIconPosition(icons, 46, 0, 100);
-			// then
-			expect(result).toBe(2);
-		});
-
-		it('should return -1 when value is null', () => {
-			// given
-			const icons = ['icon1', 'icon2', 'icon3', 'icon4', 'icon5'];
-			// when
-			const result = getSelectedIconPosition(icons, null, 0, 100);
-			// then
-			expect(result).toBe(-1);
 		});
 	});
 });
